DSND-1580: Updated Appointments List 404 when service unavailable
A 404 will be received from an API client under two conditions
   1. The requested resource was not present in the database, or
   2. The target service is not available
When the services are hosted on ECS the latter case will be fixed and a 500 response will
be used to signal a service is down.
In the meantime, a zero length response body is seen when a resource is not found (case 1),
and a non-zero response body is seen when a service is unavailable (case 2).